--- configurations/hardhat/package.json
@@ -1,58 +1,0 @@
-{
-  "name": "@lightdotso/hardhat-config",
-  "version": "1.0.0",
-  "private": false,
-  "homepage": "https://github.com/LightDotSo/LightDotSo",
-  "bugs": {
-    "url": "https://github.com/LightDotSo/LightDotSo/issues"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/LightDotSo/LightDotSo.git"
-  },
-  "license": "GPL-3.0",
-  "author": "Light, Inc.",
-  "main": "dist/index.js",
-  "module": "dist/index.mjs",
-  "source": "src/index.ts",
-  "typings": "dist/index.d.ts",
-  "files": [
-    "src/index.js"
-  ],
-  "scripts": {
-    "build": "pnpm run clean && pnpm run compile",
-    "clean": "rimraf -rf ./dist",
-    "compile": "tsc",
-    "prepublishOnly": "pnpm run build"
-  },
-  "dependencies": {
-    "@ethersproject/abi": "^5.7.0",
-    "@ethersproject/providers": "^5.7.0",
-    "@lightdotso/tsconfig": "workspace:*",
-    "@matterlabs/hardhat-zksync-deploy": "^0.4.0",
-    "@matterlabs/hardhat-zksync-solc": "^0.3.7",
-    "@nomiclabs/hardhat-ethers": "^2.1.1",
-    "@nomiclabs/hardhat-etherscan": "^3.1.0",
-    "@nomiclabs/hardhat-waffle": "^2.0.3",
-    "@openzeppelin/hardhat-upgrades": "^1.20.0",
-    "@typechain/ethers-v5": "^10.1.0",
-    "@typechain/hardhat": "^6.1.2",
-    "@types/chai": "^4.3.3",
-    "@types/jest": "^29.0.0",
-    "chai": "^4.3.6",
-    "dotenv": "^16.0.2",
-    "ethers": "^5.7.0",
-    "hardhat": "^2.10.2",
-    "hardhat-abi-exporter": "^2.10.0",
-    "hardhat-deploy": "^0.11.12",
-    "hardhat-gas-reporter": "^1.0.8",
-    "hardhat-preprocessor": "^0.1.4",
-    "hardhat-watcher": "^2.5.0",
-    "toml": "^3.0.0",
-    "tsconfig-paths": "^4.1.0"
-  },
-  "publishConfig": {
-    "access": "public",
-    "registry": "https://registry.npmjs.org"
-  }
-}
