--- configurations/hardhat/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "compilerOptions": {
-    "outDir": "./dist"
-  },
-  "exclude": ["dist", "build", "node_modules"],
-  "extends": "@lightdotso/tsconfig/node.json",
-  "include": ["."]
-}
