--- packages/contracts/.depcheckrc.yml
@@ -1,2 +1,0 @@
-ignores:
-  - "@lightdotso/hardhat-config"
